tests/requirements.txt
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.1.0
pytest-mock==3.12.0
aiohttp==3.9.3
discord.py==2.3.2
python-dotenv==1.0.0 